Collect blog route/url methods in single trait
